requirements.txt
# MCP Framework
fastmcp>=0.1.0

# Vector Database
qdrant-client>=1.9.0
fastembed>=0.7.0

# GitHub Integration
PyGithub>=2.3.0

# Code Parsing (AST)
tree-sitter>=0.22.0
tree-sitter-python>=0.21.0
tree-sitter-java>=0.21.0
tree-sitter-javascript>=0.21.0

# LLM Provider (Gemini - new SDK)
google-genai>=1.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
httpx>=0.27.0
pydantic>=2.0.0

# Development/Testing
pytest>=8.0.0
pytest-cov>=4.1.0

